+++ bootrom/spitest/myprintf.h
@@ -6,8 +6,8 @@
 #include "uart.h"
 
 // extern void printf(char*, ...);				
-extern void printf(uint8_t *, uint64_t);
-void printNum(uint64_t); 
+extern void printf(uint8_t *, uint32_t);
+void printNum(uint32_t); 
 void putch(uint8_t);
 
 #endif
